Update embedthumbnail.py
When a thumbnail is embedded by AtomicParsley, the last access and modification time is changed to the current time. The proposed change restores the last access and modification time so that the new file, which includes the thumbnail, has the same timestamps as the original file. Since the timestamp of the original file corresponds to the upload date and time of the file, this is important if you rely on this timestamp.